move calendar stuff (team-g) to own folders (cleanup)
