+++ src/pages/Calendar.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { NavLink, Link, Outlet } from "react-router"
-import '../styles/Calendar.css'
-import CalendarHeader from '../components/CalendarHeader.jsx'
+import '../styles/calendar/Calendar.css'
+import CalendarHeader from '../components/calendar/CalendarHeader.jsx'
 
 function Calendar() {
   return (
